Multiple schedules for subjects in multiple semesters
Add new attributes scheduleFall, scheduleIAP, and scheduleSpring, and
store the fall, IAP, and/or spring schedules of a subject here if the
subject is offered and has schedules available in more than one
semester.

In the case of such subjects, the schedule attribute is kept and
populated with one of these schedules (for backwards compatibility), but
the specific semester that is chosen is undefined. (Client code should
check for the presence of scheduleFall/scheduleIAP/scheduleSpring --
depending on the semester -- and prefer the value of that attribute to
that of the schedule attribute).