Propagate old ID from previous semesters
The fall 2023 catalog no longer contains the old IDs of courses that
were renumbered in 2022; to aid people who are used to the old
numbering, propagate the old ID column from previous semesters to the
consensus catalog, even if it is not present for a subject in the most
recent catalog.

This commit also includes some fixes to consensus_catalog.py so that it
can be run as a standalone script for testing purposes.